separated adventurers component from adventurersPage and added no adventurer component
